apps/backend/pyproject.toml
[project]
name = "j-feelgood-subscriptions"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "read.musik", email = "read.musik@gmail.com"},
]
dependencies = [
    "uvicorn>=0.35.0",
    "fastapi>=0.116.1",
    "sqlmodel>=0.0.24",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "bcrypt>=4.3.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "stripe>=12.4.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
